Extension activation cleanup
